Adding the FTE calculation (#296)
* Adding the FTE calculation

Per user request. This change includes a new column in the scenarios tab, additions to the analysis pipeline, and updates to documentation.

* Update model_inputs_demo.xlsx

Each scenario requires a set of specification in "CadreRoles" sheet. Added cadre roles specifications for scenario "StepCoverage".

* Revert "Update model_inputs_demo.xlsx"

This reverts commit bbb79f3eeb3aae0392ea6bf7f9ce72c128c3b1d2.

* Update model_inputs_demo.xlsx

Added cadre roles for the scenario "StepCoverage". Each scenario is required to have a set of specified cadre roles in the "CadreRoles" sheet.

* Update Analysis_pipeline_demo

Updated labels on FTE plots to include MaxUtilization values.

* Update minor version

update minor version as we introduce a input sheet change for FTE analysis which is backward compatible.

---------

Co-authored-by: Rui Han <rui.han@gatesfoundation.org>
Co-authored-by: MeWu-IDM <40474438+MeWu-IDM@users.noreply.github.com>